refactor(70559): Refatora alguns testes
